requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pyarrow>=12.0.0
google-genai
pydantic>=2.0.0
